requirements.txt
fastapi[standard]
pydantic
python-dotenv
langchain_community
google-api-python-client
langchain-google-vertexai
# deeplake